Float: Added Code generation
Added assembly generation for loading and storing floats as well as
passing floats in xmmX registers to functions (important for debugging).
It is still missign a lot of code gen stuff, but a simple addition
was working.

Also added the capability to expressions to track the resulting data
type, which is very important for deciding whether to run code
generation in regular mode or float mode. This is still missing a lot,
most importantly, I need to make type casts next, before I can implement
type tracking in the parser properly. For now everything is just
hardcoded to be integer type, so old behaviour is not broken.