Test: Corrected recursion error in unit test
Variables can not be defined by themselves or functioncalls with
themselves as arguments, unless they are already initialized.
This unit test contained just such a case, where a variable was
initialized with the result of a function call with itself as an
argument.
Changes that, so that it reliably tests, what is supposed to be tested.