Types: Automatic detection of operand types
Binary operations now detect the type of their operands, so they
can generate implicit casts later. This is crucial for working
with integers and floats together.

Also intrduced a binsry cast matrix, which determines what combination
of types will result in what type. Currently for most expression this is
filled with placeholder information.