Types: Renamed float to real.
Renamed float to real since that is what those numbers are supposed to
be in the first place. It does not matter wether or not they have a
floating point or a fixed point or no point at all. They are
representing the closest thing a computer knows about real numbers and
thus they are called real now.