IC: Intermediate Code Generation
Introduced a step between the parsing/ast and the assembly generation:
intermediate code. This is just a stub for now and is ignored because of
incomplete implementation, but will be important later for optimization
passes and is quite handy for calculating memory addresses.

In future, it will bear most of the recursive parsing of the AST, and
the assembly generation will only have to deal with a simple list of
ICs. This will make it easier to write assembly generation code and
therefore make it feasable to target multiple platforms.

The IC stage is also supposed to simplify the input to the assembly
generation stage, such that not too much decisions have to be taken
there.

The IC stage is implemented as linked list of four-address-codes.