Parse: Correct parsing of semicolons
Made sure semicolons are only removed from the token list in places
specieifed in the Backus-Naur form. This is critical for other parts of
the parsing to work correctly since some expressions can contain
themselves.

Also began correct implementation of the postfix expression parsing
which is neccessary for dealing with arrays and struct/union member
access.