IC Gen: Detect external functions
If a function is not found in the symbol list, atm the compiler has to
trust the programmer that it will be linked later. In the future, this
should flag the function call and emit an error on linking, but for now
that is not possible since the compiler relies on an external linker.
However we still need to distinguish external functions from those
within the translation unit since internal ones need to be connected to
the right symbol.