asm_gen/ic_gen/symbols: dedicated varaibles for storage locations
Added dedicated variables for storage locations to symbols.
This enabled making a more intelligent register selection function
and therefore enabled a lot of asm code to be correct now.

There are still a lot of hickups, since the register selector does not
yet decide on register reuse, but this could also be a problem with
incomplete ic generation.